management join between appUser and appUserRole

+++ src/main/java/me/robnette/spring_angular/controller/AppUserRestController.java
@@ -2,6 +2,9 @@ package me.robnette.spring_angular.controller;
 
 import me.robnette.spring_angular.dao.AppUser;
 import me.robnette.spring_angular.exception.ForbiddenException;
+import me.robnette.spring_angular.model.UserPojo;
+import me.robnette.spring_angular.security.SecurityUser;
+import me.robnette.spring_angular.service.AppUserService;
 import me.robnette.spring_angular.util.Util;
 import me.robnette.spring_angular.repository.AppUserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,20 +22,20 @@ import java.util.List;
 public class AppUserRestController {
 	@Autowired
 	private AppUserRepository appUserRepository;
+	@Autowired
+	private AppUserService appUserService;
 
 	@PreAuthorize("hasRole('ROLE_ADMIN')")
 	@RequestMapping(value = "/users", method = RequestMethod.GET)
 	public List<AppUser> users() {
-//		SecurityUser securityUser = (SecurityUser)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-//		System.out.println(securityUser);
 		return appUserRepository.findAll();
 	}
 
 
 	@PreAuthorize("hasRole('ROLE_ADMIN')")
-	@RequestMapping(value = "/users/{id}", method = RequestMethod.GET)
-	public ResponseEntity<AppUser> userById(@PathVariable Long id) {
-		AppUser appUser = appUserRepository.findOne(id);
+	@RequestMapping(value = "/users/{uid}", method = RequestMethod.GET)
+	public ResponseEntity<AppUser> userByUid(@PathVariable String uid) {
+		AppUser appUser = appUserRepository.findOneByUid(uid);
 		if (appUser == null) {
 			return new ResponseEntity<AppUser>(HttpStatus.NO_CONTENT);
 		} else {
@@ -42,15 +45,16 @@ public class AppUserRestController {
 
 
 	@PreAuthorize("hasRole('ROLE_ADMIN')")
-	@RequestMapping(value = "/users/{id}", method = RequestMethod.DELETE)
-	public ResponseEntity<String> deleteUser(@PathVariable Long id) {
-		AppUser appUser = appUserRepository.findOne(id);
-		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-		String loggedUsername = auth.getName();
+	@RequestMapping(value = "/users/{uid}", method = RequestMethod.DELETE)
+	public ResponseEntity<String> deleteUser(@PathVariable String uid) {
+		SecurityUser securityUser = (SecurityUser)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		if (securityUser.getUid().equals(uid)) {
+			throw new ForbiddenException("You cannot delete your account");
+		}
+
+		AppUser appUser = appUserRepository.findOneByUid(uid);
 		if (appUser == null) {
 			return new ResponseEntity<String>(HttpStatus.NO_CONTENT);
-		} else if (appUser.getUsername().equalsIgnoreCase(loggedUsername)) {
-			throw new ForbiddenException("You cannot delete your account");
 		} else {
 			appUserRepository.delete(appUser);
 			return new ResponseEntity<String>(HttpStatus.OK);
@@ -60,33 +64,27 @@ public class AppUserRestController {
 
 	@PreAuthorize("hasRole('ROLE_ADMIN')")
 	@RequestMapping(value = "/users", method = RequestMethod.POST)
-	public ResponseEntity<AppUser> createUser(@RequestBody AppUser appUser) {
-		if (appUserRepository.findOneByUsername(appUser.getUsername()) != null) {
-			throw new ForbiddenException("Username already exist");
-		}
-
-		Util.passwordCheck(appUser.getPassword());
-
-		appUser.setPassword(Util.passwordEncode(appUser.getPassword()));
-		appUser.setUid(Util.getUid());
-
+	public ResponseEntity<AppUser> createUser(@RequestBody UserPojo userPojo) {
+		AppUser appUser = appUserService.addUser(userPojo);
 		return new ResponseEntity<AppUser>(appUserRepository.save(appUser), HttpStatus.CREATED);
 	}
 
 
 	@PreAuthorize("hasRole('ROLE_ADMIN')")
 	@RequestMapping(value = "/users", method = RequestMethod.PUT)
-	public AppUser updateUser(@RequestBody AppUser appUser) {
-		if (appUserRepository.findOneByUsername(appUser.getUsername()) != null
-				&& appUserRepository.findOneByUsername(appUser.getUsername()).getId() != appUser.getId()) {
-			throw new ForbiddenException("Username already exist");
-		}
+	public AppUser updateUser(@RequestBody UserPojo userPojo) {
+		AppUser appUser = appUserService.updateUser(userPojo);
 
-		Util.passwordCheck(appUser.getPassword());
-
-		appUser.setPassword(Util.passwordEncode(appUser.getPassword()));
-
-		return appUserRepository.save(appUser);
+//		AppUser appUser = appUserRepository.findOneByUid(userPojo.getUid());
+//		if(appUser == null){
+//			throw new ForbiddenException("User unknown");
+//		}
+//		if (! appUser.getUsername().equals(userPojo.getUsername()) && appUserRepository.findOneByUsername(userPojo.getUsername()) != null) {
+//			throw new ForbiddenException("Username already exist");
+//		}
+//		appUser.updateUser(userPojo);
+//		return appUserRepository.save(appUser);
+		return appUser;
 	}
 
 }
